fix(usage): infer the account filter type through the provider handlers
`unknown` for accountFilter broke parameter contravariance against each
provider store, and the slice file passed its max-lines budget once the
account-scoped state landed; split the factory out.
