+++ src/main/ipc/usage-provider-handlers.ts
@@ -11,36 +11,43 @@ type UsageProviderStores = {
 
 type UsageProviderChannelPrefix = keyof UsageProviderStores
 
-type UsageProviderHandlerStore<Scope, Range, BreakdownKind> = {
+// Why: AccountFilter is inferred from each store so a provider-specific filter
+// type stays assignable — `unknown` here would break parameter contravariance.
+type UsageProviderHandlerStore<Scope, Range, BreakdownKind, AccountFilter> = {
   getScanState: () => unknown
   setEnabled: (enabled: boolean) => unknown
   refresh: (force?: boolean) => unknown
-  getSnapshot: (scope: Scope, range: Range, limit?: number, accountFilter?: unknown) => unknown
-  getSummary: (scope: Scope, range: Range, accountFilter?: unknown) => unknown
-  getDaily: (scope: Scope, range: Range, accountFilter?: unknown) => unknown
+  getSnapshot: (
+    scope: Scope,
+    range: Range,
+    limit?: number,
+    accountFilter?: AccountFilter
+  ) => unknown
+  getSummary: (scope: Scope, range: Range, accountFilter?: AccountFilter) => unknown
+  getDaily: (scope: Scope, range: Range, accountFilter?: AccountFilter) => unknown
   getBreakdown: (
     scope: Scope,
     range: Range,
     kind: BreakdownKind,
-    accountFilter?: unknown
+    accountFilter?: AccountFilter
   ) => unknown
   getRecentSessions: (
     scope: Scope,
     range: Range,
     limit?: number,
-    accountFilter?: unknown
+    accountFilter?: AccountFilter
   ) => unknown
 }
 
-type UsageRangeArgs<Scope, Range> = {
+type UsageRangeArgs<Scope, Range, AccountFilter> = {
   scope: Scope
   range: Range
-  accountFilter?: unknown
+  accountFilter?: AccountFilter
 }
 
-function registerProviderHandlers<Scope, Range, BreakdownKind>(
+function registerProviderHandlers<Scope, Range, BreakdownKind, AccountFilter>(
   prefix: UsageProviderChannelPrefix,
-  usage: UsageProviderHandlerStore<Scope, Range, BreakdownKind>
+  usage: UsageProviderHandlerStore<Scope, Range, BreakdownKind, AccountFilter>
 ): void {
   ipcMain.handle(`${prefix}:getScanState`, () => usage.getScanState())
   ipcMain.handle(`${prefix}:setEnabled`, (_event, args: { enabled: boolean }) =>
@@ -51,23 +58,27 @@ function registerProviderHandlers<Scope, Range, BreakdownKind>(
   )
   ipcMain.handle(
     `${prefix}:getSnapshot`,
-    (_event, args: UsageRangeArgs<Scope, Range> & { limit?: number }) =>
+    (_event, args: UsageRangeArgs<Scope, Range, AccountFilter> & { limit?: number }) =>
       usage.getSnapshot(args.scope, args.range, args.limit, args.accountFilter)
   )
-  ipcMain.handle(`${prefix}:getSummary`, (_event, args: UsageRangeArgs<Scope, Range>) =>
-    usage.getSummary(args.scope, args.range, args.accountFilter)
+  ipcMain.handle(
+    `${prefix}:getSummary`,
+    (_event, args: UsageRangeArgs<Scope, Range, AccountFilter>) =>
+      usage.getSummary(args.scope, args.range, args.accountFilter)
   )
-  ipcMain.handle(`${prefix}:getDaily`, (_event, args: UsageRangeArgs<Scope, Range>) =>
-    usage.getDaily(args.scope, args.range, args.accountFilter)
+  ipcMain.handle(
+    `${prefix}:getDaily`,
+    (_event, args: UsageRangeArgs<Scope, Range, AccountFilter>) =>
+      usage.getDaily(args.scope, args.range, args.accountFilter)
   )
   ipcMain.handle(
     `${prefix}:getBreakdown`,
-    (_event, args: UsageRangeArgs<Scope, Range> & { kind: BreakdownKind }) =>
+    (_event, args: UsageRangeArgs<Scope, Range, AccountFilter> & { kind: BreakdownKind }) =>
       usage.getBreakdown(args.scope, args.range, args.kind, args.accountFilter)
   )
   ipcMain.handle(
     `${prefix}:getRecentSessions`,
-    (_event, args: UsageRangeArgs<Scope, Range> & { limit?: number }) =>
+    (_event, args: UsageRangeArgs<Scope, Range, AccountFilter> & { limit?: number }) =>
       usage.getRecentSessions(args.scope, args.range, args.limit, args.accountFilter)
   )
 }
